More readability stuff
I'm not sure how some of these words are indexed.  For instance, book is listed as a recognized word, but notebook is not.  Do we thing that makes it harder to read?  What about a case like destroy?  Destroy is in there, but destroyed is not...idk how these tests treat cases.