Thoughts about topics
So I think we can shorten the base education section to the three questions posed in the top of what I added.  The ELI5 version is all they should have to see at first, and I think we can adjust the check vs CC example accordingly.  If the reader is interested in learning more, they can dive deeper into whatever it is that interests them; a breakdown of how i think those should be grouped is also included.  I think when we introduce the new concepts, we should re-present the check vs CC example with the updated level of detail